Replace isInt() with isNumber() and get/asInt() with asDouble(), sometimes static_cast back to int/int64_t (#2561)

+++ vnext/ReactUWP/Views/PickerViewManager.cpp
@@ -115,7 +115,7 @@ void PickerShadowNode::updateProperties(const folly::dynamic&& props)
     {
       if (propertyValue.isNumber())
       {
-        auto selectedIndex = propertyValue.asInt();
+        auto selectedIndex = propertyValue.asDouble();
         if (selectedIndex == static_cast<int32_t>(selectedIndex))
         {
           m_selectedIndex = static_cast<int32_t>(selectedIndex);
